Rename io signals and remove top register
`Top` register is not needed to compare to `top_in`.
Instead we handle this with a `clr` signal, which resets the counter
when pulled to `High`.

We also added `setName` after all the input signals, so they have a
non-autogenerated name in verilog compile and simulation.

This commit also adds the compiled verilog code of the SpinalHDL Modcounter