Make PWM and FreqDiv compile
This commit contains a lot of changes, but almost no logic changes.

All these changes are made to get the code to compile.
The reason for not compiling in the end came to me creating a bad
build.sbt file. So it was not even the codes fault.

BUT TLDR:

This version builds and is stable.