FreqDiv: Adjust INT calculation to fit RP2350 specsheet
To fit the specsheet of the RP2350, which describes this:

"Set DIV_INT to 0 to divide the count rate by the maximum possible value
of 256. You must not set any DIV_FRAC bits when DIV_INT is 0" (see RP2350 datasheet, page 1084)

We have to use `div_in - 1` for our comparison with `ctr`.