Fix modulo operation resetting to base operation
The modulo operation should not reset to base value, instead keep on
adding up with the modulo applied.

We fix this by using a combinational logic and assigning it to `frac_acc`
when we have a modulo overflow.

IDEAS: The SpinalHDL syntax probably has a cleaner option for this, but
I could not yet find it.